requirements.txt
aiosqlite~=0.20.0
matplotlib~=3.9.1
aiohttp~=3.9.1
pillow~=10.2.0
discord.py>=2.3.2
aiofiles